do not use a GRAPE type here
git-svn-id: https://dune.mathematik.uni-freiburg.de/svn/dune-fem/trunk@4048 4a6c9bf5-c326-0410-8cc6-c7cabc8ddcf0

+++ fem/io/visual/grape/datadisp/readioparams.cc
@@ -125,7 +125,7 @@ inline int readParameterList (int argc, char **argv, bool displayData = true )
 
   int    i_delta = 1;
   const  char *replay = 0;
-  REAL   timestep = 1.0e-3;
+  double   timestep = 1.0e-3;
   bool fixedMesh = false;
   
   info = (INFO *) malloc(n_info*sizeof(INFO));
